Prepend protocol for actual follow, and strip for label
